+++ src/main/java/com/core/reactive/corereactive/util/api/object/JsonCommunityDragon.java
@@ -2,8 +2,6 @@ package com.core.reactive.corereactive.util.api.object;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -11,8 +9,8 @@ import java.math.BigDecimal;
 @Builder
 @ToString
 public class JsonCommunityDragon {
-    private BigDecimal attackSpeed;
-    private BigDecimal gameplayRadius;
-    private BigDecimal windUp;
-    private BigDecimal windupMod;
+    private Double attackSpeed;
+    private Double gameplayRadius;
+    private Double windUp;
+    private Double windupMod;
 }
